backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.42
asyncpg==0.30.0
pydantic-settings==2.10.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.20
pillow==11.3.0
pillow-heif==0.22.0
